insights: refactor and cleanup
